fix: select component on click, hide private fns
